+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to the electron-react-app (ERA) are listed here.
 - Update layout to new design
 - Self host auth0 callback page [(#58)](https://github.com/heyito/ito/pull/58)
 - Better handle errors from transcription service [(#62)](https://github.com/heyito/ito/pull/62)
+- Add context to the transcription + improving prompting [(#64)](https://github.com/heyito/ito/pull/64)
 
 ### v0.2.2